pyproject.toml
[tool.poetry]
name = "yalom"
version = "0.1.0"
description = "YouTube transcript processing project"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "yalom"}]

[tool.poetry.dependencies]
python = "^3.10"
youtube-transcript-api = "^0.6.2"
yt-dlp = "^2024.0.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"
alembic = "^1.13.0"
pandas = "^2.0.0"
langchain = "^1.2.6"
langchain-openai = "^1.1.7"
langchain-chroma = "^1.1.0"
qdrant-client = "^1.16.2"
langchain-qdrant = "^1.1.0"
langchain-huggingface = "^1.2.0"
langchain-community = "^0.4.1"
langchain-ollama = "^1.0.1"
langchain-groq = "^1.0.0"
streamlit = "^1.53.1"
langdetect = "^1.0.9"
boto3 = "^1.35.0"
pinecone-client = "^5.0.1"
openai = "^1.59.9"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
ipykernel = "^6.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
